% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctree_input}
\alias{ctree_input}
\title{Example ctree dataset.}
\format{
A list of parameters to build a clone tree,
}
\usage{
data(ctree_input)
}
\description{
Example ctree dataset.
}
\examples{
data(ctree_input)
ctree_input

}
\keyword{datasets}
