% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clone_size.R
\name{plot_clone_size}
\alias{plot_clone_size}
\title{Plot a clone size histogram, and test for.}
\usage{
plot_clone_size(x, alpha_level = 0.05)
}
\arguments{
\item{x}{A \code{ctree} tree.}

\item{alpha_level}{Alpha level for the test, default is 0.05.}
}
\value{
A \code{ggplot} plot.
}
\description{
This function creates a \code{ggplot}
barplot of the clone size values foe each clone in the patient's data.
The size of a clone is defined as the number of mutations assigned 
to it, and is provided in input. 

The barplot is annotated to report wether a subclone with a driver
is significantly larger than the expected size for a subclone without
driver. To carry out this test subclones without drivers are used to
estimate the parameters of a univariate Gaussian distribution (mean
and standard deviation), the p-value is then computed from the fit
distribution through the `pnorm` function. 

The confidence level for the test can be passed as parameter.
}
\examples{
data('ctree_input')

x = ctrees(
   ctree_input$CCF_clusters,
   ctree_input$drivers,
   ctree_input$samples,
   ctree_input$patient,
   ctree_input$sspace.cutoff,
   ctree_input$n.sampling,
   ctree_input$store.max
   )
   
plot_clone_size(x[[1]])
}
