% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_information_transfer.R
\name{plot_information_transfer}
\alias{plot_information_transfer}
\title{Plot the information transfer of a tree.}
\usage{
plot_information_transfer(x,
  node_palette = colorRampPalette(RColorBrewer::brewer.pal(n = 9,
  "Set1")), tree_layout = "tree", ...)
}
\arguments{
\item{x}{A \code{ctree} tree.}

\item{node_palette}{A function that can return, for an input number,
a number of colours.}

\item{tree_layout}{Layout of this model, as of \code{ggraph}.}

\item{...}{Other parameters, not used in this case.}
}
\value{
A \code{ggplot} object for the plot
}
\description{
The information transfer of the tree is the set of orderings
associated to the internal annotated driver events.

This function plots these orderings following a topological
sort of the node of the corresponding clone tree.
}
\examples{
data(ctree_input)

x = ctrees(
ctree_input$CCF_clusters, 
ctree_input$drivers,
ctree_input$samples,
ctree_input$patient,
ctree_input$sspace.cutoff,
ctree_input$n.sampling,
ctree_input$store.max
)

plot_information_transfer(x[[1]])

# Change layout -- use igraph's "kk" layout
plot_information_transfer(x[[1]], tree_layout = 'kk')
}
