% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.ctree}
\alias{print.ctree}
\title{Print a \code{"rev_phylo"} object.}
\usage{
\method{print}{ctree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rev_phylo"}.}

\item{...}{Extra parameters}
}
\value{
Nothing
}
\description{
Print a \code{"rev_phylo"} object.
}
\details{
Print a summary for a \code{"rev_phylo"} object, which includes MSDOS-like console layout for trees.
}
\examples{
data(CRC.cohort)
CRC.cohort$phylogenies[['adenoma_3']][[1]]
}
