% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.ctree}
\alias{summary.ctree}
\title{Summary of a \code{"ctree"} object.}
\usage{
\method{summary}{ctree}(x, ...)
}
\arguments{
\item{x}{A \code{ctree} tree.}

\item{...}{Extra parameters}
}
\value{
None.
}
\description{
Reports some summary statistics for a \code{"rev_phylo"} object, 
which is a bit more than just using \code{print.}
}
\examples{
data(ctree_input)

x = ctrees(
ctree_input$CCF_clusters, 
ctree_input$drivers,
ctree_input$samples,
ctree_input$patient,
ctree_input$sspace.cutoff,
ctree_input$n.sampling,
ctree_input$store.max
)

summary(x[[1]])
}
