% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.ctree}
\alias{print.ctree}
\title{Print a \code{ctree} tree.}
\usage{
\method{print}{ctree}(x, ...)
}
\arguments{
\item{x}{A \code{ctree} tree.}

\item{...}{Extra S3 parameters}
}
\value{
Nothing
}
\description{
Print a \code{ctree} tree.
}
\details{
Print a summary for a \code{ctree} object, which includes a
shell-frinedly layout and other information.
}
\examples{
data(ctree_input)

x = ctrees(
ctree_input$CCF_clusters, 
ctree_input$drivers,
ctree_input$samples,
ctree_input$patient,
ctree_input$sspace.cutoff,
ctree_input$n.sampling,
ctree_input$store.max
)

print(x[[1]])
}
