% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.ctree}
\alias{print.ctree}
\title{Print a \code{ctree} tree.}
\usage{
\method{print}{ctree}(x, ...)
}
\arguments{
\item{x}{A \code{ctree} tree.}

\item{...}{Extra S3 parameters}
}
\value{
Nothing
}
\description{
Print a \code{ctree} tree.
}
\details{
Print a summary for a \code{ctree} object, which includes a
shell-frinedly layout and other information.
}
\examples{
data(CRC.cohort)
CRC.cohort$phylogenies[['adenoma_3']][[1]]
}
